Move to a module-based approach, add a dedicated run script
