+++ pitrainer/run.py
@@ -1,0 +1,5 @@
+#!/usr/bin/env python3
+from pitrainer import pitrainer
+
+if __name__ == "__main__":
+    pitrainer()
